tp_app/pyproject.toml
[project]
name = "tp-app"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "streamlit==1.41.1",
    "langchain>=0.3.0",
    "langchain-groq>=0.2.0",
    "langchain-openai>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-docling>=0.1.0",
    "langgraph>=0.2.0",
    "python-docx>=1.1.0",
    "tavily-python>=0.5.0",
    "docling>=2.0.0",
    "faiss-cpu>=1.8.0",
    "sentence-transformers>=3.0.0",
    "pageindex>=0.2.5.1",
    "pymupdf>=1.27.1",
    "pypdf2>=3.0.1",
    "tiktoken>=0.12.0",
    "pyyaml>=6.0.3",
    "openpyxl>=3.1.5",
    "langchain-huggingface>=1.2.0",
    "pywin32>=311",
    "docxtpl>=0.20.2",
]
